+++ tests/core/utils/testSetup.js
@@ -1,8 +1,6 @@
 import { configure } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-
-/* eslint-disable */
-global.console = { log: jest.fn() };
+import "jest-location-mock";
 
 /* eslint-disable */
 global.navigator.clipboard = {