+++ tests/test-suites/components/AppModal.tsx
@@ -1,13 +1,14 @@
-import { initStore, openModal } from "@cianciarusocataldo/modular-engine";
-import { mount } from "enzyme";
-import React, { Suspense } from "react";
+import React from "react";
+import { Store } from "redux";
+
 import { Provider } from "react-redux";
+import { mount } from "enzyme";
+
+import { openModal } from "@cianciarusocataldo/modular-engine";
+
 import AppModal from "../../../src/app/components/AppModal";
 
-const AppModalTest = () => {
-  const { store } = initStore({
-    config: { redux: { ui: true, modal: true } },
-  });
+const AppModalTest = (store: Store) => {
   describe("\n     AppModal\n", () => {
     test("renders correctly", () => {
       let wrapper = mount(
@@ -15,20 +16,36 @@ const AppModalTest = () => {
           <AppModal modals={{}} />
         </Provider>
       );
-      expect(wrapper);
+      expect(wrapper.find("#modular-modal").length).toBeGreaterThan(0);
+    });
+    test("after dispatching openModal action, shows the correct form", () => {
       store.dispatch(openModal("test"));
-      wrapper = mount(
-        <Suspense fallback={<div />}>
-          <Provider store={store}>
-            <div>
-              <AppModal modals={{ test: () => <div>test form</div> }} />
-            </div>
-            <div>test div</div>
-          </Provider>
-        </Suspense>
+
+      const wrapper = mount(
+        <Provider store={store}>
+          <AppModal
+            modals={{
+              test: () => <div data-id="test-form">test form</div>,
+            }}
+          />
+        </Provider>
+      );
+      expect(wrapper.find('div[data-id="test-form"]').length).toBeGreaterThan(
+        0
+      );
+    });
+
+    test("close correctly when closeModal action is dispatched", () => {
+      const wrapper = mount(
+        <Provider store={store}>
+          <AppModal
+            modals={{
+              test: () => <div data-id="test-form">test form</div>,
+            }}
+          />
+        </Provider>
       );
       wrapper.find(".close-button").at(1).simulate("click");
-      expect(wrapper);
     });
   });
 };