+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+# 0.7.0
+
+- Custom `onStart` callback, called at the start of the init process
+
 # 0.6.0
 
 - Configs as parameters (if not set, default standard predefined files will be used)
 - Types updated
-- Statically importing modular-engine instead of promise
+- Statically importing `modular-engine` instead of promise
 
 # 0.5.0
 
@@ -14,14 +18,14 @@
 
 - Modular-engine updated
 - Clean up
-- Error fallback customizable through config.error field
+- Error `fallback` customizable through config.error field
 
 # 0.3.0
 
 - Theme actions updated
 - Types definition
 - Docs
-- Error boundary for the whole app, with customizable fallback
+- `Error boundary` for the whole app, with customizable fallback
 
 # 0.2.0